Move common functionality to a trait.
